Split simulated sled agent into sub-modules
